added badge to notification; added visibility options for notifications; renamed service
