+++ src/services/1inch.execute.tsx
@@ -5,7 +5,7 @@ import { oneInch } from './1Inch'
 import { createAndPushProcess, initStatus, setStatusDone, setStatusFailed } from './status'
 import { getApproved, setApproval } from './utils'
 import { constants } from 'ethers';
-import localNotifications, { NotificationType } from './localNotifications'
+import notifications, { NotificationType } from './notifications'
 
 export const executeOneInchSwap = async (chainId: number, signer: JsonRpcSigner, srcToken: string, destToken: string, srcAmount: BigNumber, srcAddress: string, destAddress: string, updateStatus?: Function, initialStatus?: Execution) => {
 
@@ -83,7 +83,7 @@ export const executeOneInchSwap = async (chainId: number, signer: JsonRpcSigner,
     if (e.message) swapProcess.errorMessage = e.message
     if (e.code) swapProcess.errorCode = e.code
     setStatusFailed(update, status, swapProcess)
-    localNotifications.showNotification(NotificationType.SWAP_ERROR)
+    notifications.showNotification(NotificationType.SWAP_ERROR)
     throw e
   }
 
@@ -95,7 +95,7 @@ export const executeOneInchSwap = async (chainId: number, signer: JsonRpcSigner,
   status.gasUsed = (status.gasUsed || 0) + parsedReceipt.gasUsed
   status.status = 'DONE'
   setStatusDone(update, status, swapProcess)
-  localNotifications.showNotification(NotificationType.SWAP_SUCCESSFUL)
+  notifications.showNotification(NotificationType.SWAP_SUCCESSFUL)
 
   // DONE
   return status
