+++ src/services/notifications.ts
@@ -7,6 +7,7 @@ export enum NotificationType {
   CROSS_ERROR
 }
 
+
 const notificationsSupported = ():boolean =>{
   if (!("Notification" in window)) {
     console.log("This browser does not support desktop notification");
@@ -24,36 +25,43 @@ const activateNotifications = async (): Promise<NotificationPermission>  => {
   return "denied"
 }
 
-const getNotificationContents = (type: NotificationType):{title: string, options: NotificationOptions} => {
+const getNotificationContents = (type: NotificationType):{title: string, options: NotificationOptions, alwaysShow: boolean} => {
   let title: string = "";
   let options: NotificationOptions = {
     dir: "auto",
-    icon: lifiIcon,
+    // icon: lifiIcon,
+    badge: lifiIcon
   }
+  let alwaysShow = false
   switch(type){
     case NotificationType.SWAP_SUCCESSFUL:
       title = "Swap Successful!"
       break;
     case NotificationType.SWAP_ERROR:
       title = "Swap Failed!"
+      alwaysShow = true
       break;
       case NotificationType.CROSS_SUCCESSFUL:
         title = "Cross Chain Transfer Successful!"
         break;
       case NotificationType.CROSS_ERROR:
         title = "Cross Chain Transfer Failed!"
+        alwaysShow = true
         break;
     default:
       break;
   }
 
-  return {title, options}
+  return {title, options, alwaysShow}
 }
 
 const showNotification = (type: NotificationType) => {
   if(Notification.permission === "denied") return
-  const {title, options} = getNotificationContents(type)
-  new Notification(title, options)
+  const {title, options, alwaysShow} = getNotificationContents(type)
+  if(document.hidden || alwaysShow){
+    new Notification(title, options)
+  }
+
 }
 
 const serviceExport = {
