+++ src/services/paraswap.execute.tsx
@@ -5,7 +5,7 @@ import { Execution, getChainById, Token } from '../types'
 import { createAndPushProcess, initStatus, setStatusDone, setStatusFailed } from './status'
 import { getApproved, setApproval } from './utils'
 import { constants } from 'ethers'
-import localNotifications, { NotificationType } from './localNotifications'
+import notifications, { NotificationType } from './notifications'
 
 export const executeParaswap = async (chainId: number, signer: JsonRpcSigner, srcToken: Token, destToken: Token, srcAmount: BigNumber, srcAddress: string, destAddress: string, updateStatus?: Function, initialStatus?: Execution) => {
 
@@ -84,7 +84,7 @@ export const executeParaswap = async (chainId: number, signer: JsonRpcSigner, sr
     if (e.message) waitingProcess.errorMessage = e.message
     if (e.code) waitingProcess.errorCode = e.code
     setStatusFailed(update, status, waitingProcess)
-    localNotifications.showNotification(NotificationType.SWAP_ERROR)
+    notifications.showNotification(NotificationType.SWAP_ERROR)
     throw e
   }
 
@@ -99,7 +99,7 @@ export const executeParaswap = async (chainId: number, signer: JsonRpcSigner, sr
   // -> set status
   status.status = 'DONE'
   update(status)
-  localNotifications.showNotification(NotificationType.SWAP_SUCCESSFUL)
+  notifications.showNotification(NotificationType.SWAP_SUCCESSFUL)
 
   // DONE
   return status
