+++ src/services/uniswaps.execute.tsx
@@ -2,7 +2,7 @@
 import { JsonRpcSigner } from '@ethersproject/providers'
 import BigNumber from 'bignumber.js'
 import { Execution, getChainById } from '../types'
-import localNotifications, { NotificationType } from './localNotifications'
+import notifications, { NotificationType } from './notifications'
 import { createAndPushProcess, initStatus, setStatusDone, setStatusFailed } from './status'
 import * as uniswap from './uniswaps'
 import { getApproved, setApproval } from './utils'
@@ -87,7 +87,7 @@ export const executeUniswap = async (chainId: number, signer: JsonRpcSigner, src
     if (e.message) allowanceProcess.errorMessage = e.message
     if (e.code) allowanceProcess.errorCode = e.code
     setStatusFailed(update, status, waitingProcess)
-    localNotifications.showNotification(NotificationType.SWAP_ERROR)
+    notifications.showNotification(NotificationType.SWAP_ERROR)
     throw e
   }
 
@@ -102,7 +102,7 @@ export const executeUniswap = async (chainId: number, signer: JsonRpcSigner, src
   // // -> set status
   status.status = 'DONE'
   update(status)
-  localNotifications.showNotification(NotificationType.SWAP_SUCCESSFUL)
+  notifications.showNotification(NotificationType.SWAP_SUCCESSFUL)
 
 
   // DONE
